+++ src/main/kotlin/org/neo4j/graphql/handler/CreateTypeHandler.kt
@@ -2,16 +2,17 @@ package org.neo4j.graphql.handler
 
 import graphql.language.Field
 import graphql.language.FieldDefinition
-import graphql.schema.idl.TypeDefinitionRegistry
+import org.neo4j.graphql.MetaProvider
 import org.neo4j.graphql.NodeFacade
 import org.neo4j.graphql.Translator
+import org.neo4j.graphql.handler.projection.ProjectionRepository
 
 class CreateTypeHandler(
         type: NodeFacade,
         fieldDefinition: FieldDefinition,
-        typeDefinitionRegistry: TypeDefinitionRegistry,
+        metaProvider: MetaProvider,
         projectionRepository: ProjectionRepository
-) : BaseDataFetcher(type, fieldDefinition, typeDefinitionRegistry, projectionRepository) {
+) : BaseDataFetcher(type, fieldDefinition, metaProvider, projectionRepository) {
 
     override fun generateCypher(variable: String, field: Field, projectionProvider: () -> Translator.Cypher, ctx: Translator.Context): Translator.Cypher {
         val properties = properties(variable, field.arguments)
