+++ src/main/kotlin/org/neo4j/graphql/handler/DeleteHandler.kt
@@ -2,19 +2,20 @@ package org.neo4j.graphql.handler
 
 import graphql.language.Field
 import graphql.language.FieldDefinition
-import graphql.schema.idl.TypeDefinitionRegistry
+import org.neo4j.graphql.MetaProvider
 import org.neo4j.graphql.NodeFacade
 import org.neo4j.graphql.Translator
+import org.neo4j.graphql.handler.projection.ProjectionRepository
 import org.neo4j.graphql.quote
 
 class DeleteHandler(
         type: NodeFacade,
         private val idField: FieldDefinition,
         fieldDefinition: FieldDefinition,
-        typeDefinitionRegistry: TypeDefinitionRegistry,
+        metaProvider: MetaProvider,
         projectionRepository: ProjectionRepository,
         val isRealtion: Boolean = type.isRelationType()
-) : BaseDataFetcher(type, fieldDefinition, typeDefinitionRegistry, projectionRepository) {
+) : BaseDataFetcher(type, fieldDefinition, metaProvider, projectionRepository) {
 
     override fun generateCypher(variable: String, field: Field, projectionProvider: () -> Translator.Cypher, ctx: Translator.Context): Translator.Cypher {
         val idArg = field.arguments.first { it.name == idField.name }
